add slice test

+++ src/NumSharp.Core/Slice.cs
@@ -45,9 +45,9 @@ namespace NumSharp
     /// </summary>
     public class Slice
     {
-        public long? Start { get; set; }
-        public long? Stop { get; set; }
-        public long Step { get; set; } = 1;
+        public int? Start { get; set; }
+        public int? Stop { get; set; }
+        public int Step { get; set; } = 1;
 
         /// <summary>
         /// Length of the slice. 
@@ -55,7 +55,7 @@ namespace NumSharp
         /// The length is not guaranteed to be known for i.e. a slice like ":". Make sure to check Start and Stop 
         /// for null before using it</remarks>
         /// </summary>
-        public long? Length => Stop - Start;
+        public int? Length => Stop - Start;
 
         /// <summary>
         /// ndarray can be indexed using slicing
@@ -64,7 +64,7 @@ namespace NumSharp
         /// <param name="start">Start index of the slice, null means from the start of the array</param>
         /// <param name="stop">Stop index (first index after end of slice), null means to the end of the array</param>
         /// <param name="step">Optional step to select every n-th element, defaults to 1</param>
-        public Slice(long? start=null, long? stop=null, long step = 1)
+        public Slice(int? start=null, int? stop=null, int step = 1)
         {
             Start = start;
             Stop = stop;
@@ -105,7 +105,7 @@ namespace NumSharp
             var single_pick_string = match.Groups[5].Value;
             if (!string.IsNullOrWhiteSpace(single_pick_string))
             {
-                if (!long.TryParse(Regex.Replace(single_pick_string ?? "", @"\s+", ""), out var start))
+                if (!int.TryParse(Regex.Replace(single_pick_string ?? "", @"\s+", ""), out var start))
                     throw new ArgumentException($"Invalid value for start: {start_string}");
                 Start = start;
                 Stop = start+1;
@@ -116,7 +116,7 @@ namespace NumSharp
                 Start = null;
             else
             {
-                if (!long.TryParse(start_string, out var start))
+                if (!int.TryParse(start_string, out var start))
                     throw new ArgumentException($"Invalid value for start: {start_string}");
                 Start = start;
             }
@@ -124,7 +124,7 @@ namespace NumSharp
                 Stop = null;
             else
             {
-                if (!long.TryParse(stop_string, out var stop))
+                if (!int.TryParse(stop_string, out var stop))
                     throw new ArgumentException($"Invalid value for start: {stop_string}");
                 Stop = stop;
             }
@@ -132,7 +132,7 @@ namespace NumSharp
                 Step = 1;
             else
             {
-                if (!long.TryParse(step_string, out var step))
+                if (!int.TryParse(step_string, out var step))
                     throw new ArgumentException($"Invalid value for start: {step_string}");
                 Step = step;
             }
@@ -177,7 +177,7 @@ namespace NumSharp
             return new Slice(null, null);
         }
 
-        public static Slice SingleValue(long index)
+        public static Slice SingleValue(int index)
         {
             return new Slice(index, index + 1);
         }
